+++ axelor-sale/src/main/java/com/axelor/apps/sale/service/SaleOrderLineServiceImpl.java
@@ -17,6 +17,15 @@
  */
 package com.axelor.apps.sale.service;
 
+import java.lang.invoke.MethodHandles;
+import java.math.BigDecimal;
+import java.math.RoundingMode;
+import java.util.HashMap;
+import java.util.Map;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import com.axelor.apps.account.db.Tax;
 import com.axelor.apps.account.db.TaxEquiv;
 import com.axelor.apps.account.db.TaxLine;
@@ -38,13 +47,6 @@ import com.axelor.exception.AxelorException;
 import com.axelor.inject.Beans;
 import com.axelor.rpc.Context;
 import com.google.inject.Inject;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.lang.invoke.MethodHandles;
-import java.math.BigDecimal;
-import java.math.RoundingMode;
-import java.util.Map;
 
 public class SaleOrderLineServiceImpl implements SaleOrderLineService {
 
@@ -95,10 +97,11 @@ public class SaleOrderLineServiceImpl implements SaleOrderLineService {
 	}
 
 	@Override
-	public BigDecimal[] computeValues(SaleOrder saleOrder, SaleOrderLine saleOrderLine) throws AxelorException {
+	public Map<String, BigDecimal> computeValues(SaleOrder saleOrder, SaleOrderLine saleOrderLine) throws AxelorException {
 
+		HashMap<String, BigDecimal> map = new HashMap<>();
 		if(saleOrder == null || (saleOrderLine.getProduct() == null && saleOrderLine.getProductName() == null) || saleOrderLine.getPrice() == null || saleOrderLine.getQty() == null)  {
-			return null;
+			return map;
 		}
 
 		BigDecimal exTaxTotal;
@@ -106,8 +109,8 @@ public class SaleOrderLineServiceImpl implements SaleOrderLineService {
 		BigDecimal inTaxTotal;
 		BigDecimal companyInTaxTotal;
 		BigDecimal priceDiscounted = this.computeDiscount(saleOrderLine);
-
 		BigDecimal taxRate = BigDecimal.ZERO;
+
 		if(saleOrderLine.getTaxLine() != null)  {  taxRate = saleOrderLine.getTaxLine().getValue();  }
 
 		if(!saleOrder.getInAti()){
@@ -127,8 +130,11 @@ public class SaleOrderLineServiceImpl implements SaleOrderLineService {
 		saleOrderLine.setExTaxTotal(exTaxTotal);
 		saleOrderLine.setCompanyInTaxTotal(companyInTaxTotal);
 		saleOrderLine.setCompanyExTaxTotal(companyExTaxTotal);
-
-		return new BigDecimal[]{exTaxTotal,inTaxTotal,companyInTaxTotal,companyExTaxTotal,priceDiscounted};
+		map.put("inTaxTotal", inTaxTotal);
+		map.put("exTaxTotal", exTaxTotal);
+		map.put("companyExTaxTotal", companyExTaxTotal);
+		map.put("companyInTaxTotal", companyInTaxTotal);
+		return map;
 	}